카드 스택을 보여주는뷰 객체 구현
